Make the majority of the API data-agnostic
