+++ scripts/normalize-data.ts
@@ -25,7 +25,7 @@ const SOURCE_RAW_DATA_PATH = resolve(
 )
 const DESTINATION_DATA_PATH = resolve(
   process.cwd(),
-  'src/app/data/products.json',
+  'src/app/data/shein-products.json',
 )
 
 const CSV_PARSER = parse({
@@ -77,9 +77,9 @@ const normalizeRecord = (record: SheinCsvRecord): SheinProduct => {
     image: (eval(images) as string[]).at(-1) || '',
 
     meta,
-    name,
+    title: name,
     price: Number(price.replace('$', '')),
-    skuId: sku.replace('SKU:', '').trim(),
+    id: sku.replace('SKU:', '').trim(),
     url,
   }
 }
@@ -94,7 +94,7 @@ const main = async () => {
     const record = rec as SheinCsvRecord
     const product = normalizeRecord(record)
 
-    products[product.skuId] = product
+    products[product.id] = product
   }
 
   await writeJson(DESTINATION_DATA_PATH, products, { spaces: 2 })
