+++ src/app/api/shein/mocks.ts
@@ -1,6 +1,6 @@
-import { ProductExtendedMessage } from '@/app/shein/types'
+import type { SheinProductExtendedMessage } from '@/app/shein/types'
 
-const MOCK_MESSAGES: ProductExtendedMessage[] = [
+export const MOCK_MESSAGES: SheinProductExtendedMessage[] = [
   {
     type: 'user',
     content:
@@ -26,7 +26,7 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
     id: 'call_pjt2bf2nztnrQofNeDgHWPWn',
     name: 'searchProducts',
     content: {
-      products: [
+      items: [
         {
           id: 'swpants07200807999',
           name: 'SHEIN BAE Zip Front Wide Waistband Coated Pants',
@@ -74,7 +74,7 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
     content:
       "Here are my top picks for you:\n\n1. `sw2211144118197961`: SHEIN PETITE High Waist Sequin Straight Leg Pants\n2. `sw2108251156521888`: SHEIN BAE High Waist PU Leather Skinny Pants\n3. `sw2209280192782655`: SHEIN Unity High Waist Stacked PU Pants\n4. `sw2209229383891421`: SHEIN BAE High Waist Glitter Wide Leg Pants\n5. `sw2209165785055175`: SHEIN BAE Solid Glitter Flare Leg Pants\n\nWith Houston's temperamental weather, these pants are perfect for the chic sparkle your dinner party fits needs! Sequins, PU leather, and glitter details provide a stylish kick, while their high waists flatter your silhouette. Some are straight, while others are wide or flare leg, offering variety for your comfort and preference. \n\nIf you wish to narrow down further, feel free to provide more specifics, like the length of the pants, your budget, or the pattern you prefer!",
     type: 'assistant',
-    skuIds: [
+    itemIds: [
       [],
       [
         'sw2211144118197961',
@@ -93,7 +93,7 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
       'If you wish to narrow down further, feel free to provide more specifics, like the length of the pants, your budget, or the pattern you prefer!',
     ],
     filter: { color: 'black', type: 'pants', style: 'party' },
-    products: {
+    items: {
       sw2211144118197961: {
         brand: 'SHEIN PETITE',
         image:
@@ -112,9 +112,9 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
           Material: 'Woven Fabric',
           Composition: 'Polyester',
         },
-        name: 'SHEIN PETITE High Waist Sequin Straight Leg Pants',
+        title: 'SHEIN PETITE High Waist Sequin Straight Leg Pants',
         price: 18.75,
-        skuId: 'sw2211144118197961',
+        id: 'sw2211144118197961',
         url: 'https://us.shein.com/SHEIN-PETITE-High-Waist-Sequin-Straight-Leg-Pants-p-12459648-cat-1740.html?mallCode=1',
       },
       sw2108251156521888: {
@@ -135,9 +135,9 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
           Material: 'PU Leather',
           Composition: 'Polyester',
         },
-        name: 'SHEIN BAE High Waist PU Leather Skinny Pants',
+        title: 'SHEIN BAE High Waist PU Leather Skinny Pants',
         price: 13.49,
-        skuId: 'sw2108251156521888',
+        id: 'sw2108251156521888',
         url: 'https://us.shein.com/SHEIN-BAE-High-Waist-PU-Leather-Skinny-Pants-p-3909051-cat-1740.html?mallCode=1',
       },
       sw2209280192782655: {
@@ -158,9 +158,9 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
           Material: 'PU Leather',
           Composition: 'Polyester',
         },
-        name: 'SHEIN Unity High Waist Stacked PU Pants',
+        title: 'SHEIN Unity High Waist Stacked PU Pants',
         price: 24.49,
-        skuId: 'sw2209280192782655',
+        id: 'sw2209280192782655',
         url: 'https://us.shein.com/SHEIN-Unity-High-Waist-Stacked-PU-Pants-p-12241226-cat-1740.html?mallCode=1',
       },
       sw2209229383891421: {
@@ -180,9 +180,9 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
           Material: 'Knitted Fabric',
           Composition: 'Polyester',
         },
-        name: 'SHEIN BAE High Waist Glitter Wide Leg Pants',
+        title: 'SHEIN BAE High Waist Glitter Wide Leg Pants',
         price: 15,
-        skuId: 'sw2209229383891421',
+        id: 'sw2209229383891421',
         url: 'https://us.shein.com/SHEIN-BAE-High-Waist-Glitter-Wide-Leg-Pants-p-12650904-cat-1740.html?mallCode=1',
       },
       sw2209165785055175: {
@@ -202,9 +202,9 @@ const MOCK_MESSAGES: ProductExtendedMessage[] = [
           Material: 'Metallic',
           Composition: 'Polyester',
         },
-        name: 'SHEIN BAE Solid Glitter Flare Leg Pants',
+        title: 'SHEIN BAE Solid Glitter Flare Leg Pants',
         price: 10.28,
-        skuId: 'sw2209165785055175',
+        id: 'sw2209165785055175',
         url: 'https://us.shein.com/SHEIN-BAE-Solid-Glitter-Flare-Leg-Pants-p-12417048-cat-1740.html?mallCode=1',
       },
     },
